PIM-5094: Rework import/export profile context to handle invalid data file
